requirements.txt
streamlit
faiss-cpu
sentence-transformers
pypdf
groq
python-dotenv
numpy
